refactor: group files for page in folder pages, the entry point is the index component.
