+++ src/components/pages/archivo/CountUp.astro
@@ -13,7 +13,7 @@ const { id, initial, final, decimals = 0, duration = 1500 } = Astro.props;
 <span id={id}></span>
 
 <script>
-  import ProgressiveNumber from "@/components/ProgressiveNumber";
+  import ProgressiveNumber from "./ProgressiveNumber";
   // @ts-ignore
   window.ProgressiveNumber = function ({
     id,
