+++ src/components/pages/archivo/Galeria.astro
