+++ src/components/pages/archivo/Index.astro
@@ -2,9 +2,9 @@
 import Layout from "@/layouts/Layout.astro"
 import Hero from "@/components/Hero.astro"
 import Button from "@/components/Button.astro"
-import Palmares from "@/components/archivo/Palmares.astro"
-import Numeros from "@/components/archivo/Numeros.astro"
-import Galeria from "@/components/archivo/Galeria.astro"
+import Palmares from "./Palmares.astro"
+import Numeros from "./Numeros.astro"
+import Galeria from "./Galeria.astro"
 
 const { i18n } = Astro.locals;
 const edicion = "2"
