+++ src/components/pages/archivo/Numeros.astro
@@ -1,6 +1,6 @@
 ---
-import CountUp from "@/components/archivo/CountUp.svelte"
-import NumerosItem from "@/components/archivo/NumerosItem.astro"
+import CountUp from "./CountUp.svelte"
+import NumerosItem from "./NumerosItem.astro"
 
 const { i18n } = Astro.props
 ---
