+++ src/components/pages/archivo/Palmares.astro
@@ -1,5 +1,5 @@
 ---
-import StreamerCard from "@/components/archivo/StreamerCard.astro"
+import StreamerCard from "./StreamerCard.astro"
 import palmares from "@/content/editions/info.json"
 
 interface Props {
