+++ src/components/pages/archivo/ProgressiveNumber.ts
