+++ src/components/pages/aviso-legal/Index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from "@/layouts/Layout.astro"
-import SubtitleItem from "@/components/legalNotice/SubtitleItem.astro"
-import DescriptionItem from "@/components/legalNotice/DescriptionItem.astro"
+import SubtitleItem from "@/components/pages/aviso-legal/SubtitleItem.astro"
+import DescriptionItem from "@/components/pages/aviso-legal/DescriptionItem.astro"
 
 const { i18n } = Astro.locals;
