+++ src/components/pages/cookies/CookieDescriptionLinkItem.astro
@@ -1,5 +1,5 @@
 ---
-import CookieLink from "@/components/cookies/CookieLink.astro"
+import CookieLink from "@/components/pages/cookies/CookieLink.astro"
 
 const { description, url } = Astro.props
 ---
