+++ src/components/pages/cookies/Index.astro
@@ -1,8 +1,8 @@
 ---
 import Layout from "@/layouts/Layout.astro";
-import CookieTypeDescriptionItem from "@/components/cookies/CookieTitleDescriptionItem.astro";
-import CookieDescriptionLinkItem from "@/components/cookies/CookieDescriptionLinkItem.astro";
-import CookieLink from "@/components/cookies/CookieLink.astro";
+import CookieTypeDescriptionItem from "@/components/pages/cookies/CookieTitleDescriptionItem.astro";
+import CookieDescriptionLinkItem from "@/components/pages/cookies/CookieDescriptionLinkItem.astro";
+import CookieLink from "@/components/pages/cookies/CookieLink.astro";
 
 const { i18n } = Astro.locals;
 const H1_STYLES =
