+++ src/components/pages/info/Index.astro
@@ -2,7 +2,7 @@
 import Layout from "@/layouts/Layout.astro"
 import Img from "@/components/Img.astro"
 import Hero from "@/components/Hero.astro"
-import InfoKitPrensa from "@/components/info/InfoKitPrensa.astro"
+import InfoKitPrensa from "@/components/pages/info/InfoKitPrensa.astro"
 
 const { i18n } = Astro.locals;
 ---
