+++ src/components/pages/info/InfoKitPrensa.astro
@@ -1,5 +1,5 @@
 ---
-import InfoKitPrensaItem from "@/components/info/InfoKitPrensaItem.astro"
+import InfoKitPrensaItem from "@/components/pages/info/InfoKitPrensaItem.astro"
 const { i18n } = Astro.locals
 ---
