+++ src/components/pages/inicio/Bento.astro
@@ -1,5 +1,5 @@
 ---
-import BentoItem from "@/components/inicio/BentoItem.astro";
+import BentoItem from "./BentoItem.astro";
 import Img from "@/components/Img.astro";
 
 const { i18n } = Astro.locals;
