+++ src/components/pages/inicio/Index.astro
@@ -1,9 +1,9 @@
 ---
 import Layout from "@/layouts/Layout.astro"
-import Video from "@/components/inicio/Video.astro"
-import Intro from "@/components/Intro.astro"
-import Bento from "@/components/inicio/Bento.astro"
-import PreFooter from "@/components/inicio/PreFooter.astro"
+import Video from "./Video.astro"
+import Intro from "./Intro.astro"
+import Bento from "./Bento.astro"
+import PreFooter from "./PreFooter.astro"
 
 const { i18n } = Astro.locals;
 ---
