+++ src/components/pages/privacidad/Index.astro
