+++ src/components/pages/vota/VoteFinal.tsx
